automation-engine/requirements.txt
# Core dependencies
flask>=3.0.0,<4.0.0
flask-restful>=0.3.10
flask-cors>=4.0.0
flask-limiter>=3.5.0
requests>=2.31.0
urllib3>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI/ML dependencies
openai>=1.0.0,<2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0

# IoT and device control
paho-mqtt>=2.0.0

# Database
sqlalchemy>=2.0.0

# Security
werkzeug>=3.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
pytest-factoryboy>=2.6.0

# Code quality
ruff>=0.3.0
mypy>=1.8.0
bandit>=1.7.7
types-requests>=2.31.0